+++ run_grievous.sh
@@ -24,8 +24,9 @@ echo "  XLerobot/Grievous Teleoperation"
 echo "======================================"
 echo ""
 echo "Select mode:"
-echo "  1) Grievous teleop of top arms without recording"
-echo "  2) Grievous teleop of top arms with recording (ZMQ streaming)"
+echo "  1) Basic XLerobot (lerobot-teleoperate, no cameras)"
+echo "  2) Grievous with cameras (remote system with cameras and ZMQ streaming)"
+echo "  3) Grievous with cameras and recording (remote system with cameras and ZMQ streaming and recording)"
 echo ""
 read -p "Enter your choice (1 or 2): " choice
 echo ""
@@ -129,6 +130,26 @@ case $choice in
         esac
         ;;
     
+    3)
+        grievous-record \
+        --robot.type=bi_so100_follower \
+        --robot.left_arm_port=$FOLLOWER_1_PORT \
+        --robot.right_arm_port=$FOLLOWER_2_PORT \
+        --robot.id=$XLEROBOT_ID \
+        --robot.cameras='{
+            left_wrist: {"type": "opencv", "index_or_path": "/dev/video6", "width": 640, "height": 480, "fps": 30},
+            right_wrist: {"type": "opencv", "index_or_path": "/dev/video8", "width": 640, "height": 480, "fps": 30},
+            head: {"type": "intelrealsense", "serial_number_or_name": "032622074046", "width": 1280, "height": 720, "fps": 30, "use_depth": true}
+        }' \
+        --teleop.type=bi_so100_leader \
+        --teleop.left_arm_port=$LEADER_1_PORT \
+        --teleop.right_arm_port=$LEADER_2_PORT \
+        --teleop.id=$LEADER_ID \
+        --dataset.repo_id=jack/test_record \
+        --dataset.num_episodes=25 \
+        --dataset.single_task="Test if recording works" \
+        --remote_client_address=tcp://*:5555
+        ;;
     *)
         echo "Invalid choice. Please run the script again and select 1 or 2."
         exit 1